% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallScale.R
\name{call.scale}
\alias{call.scale}
\title{Sigmoidal data normalization}
\usage{
call.scale(x, lambda = 2, z0 = 1.5, scaled = F)
}
\arguments{
\item{x}{array of double}

\item{lambda}{smoothness of transformation}

\item{z0}{center of sigmoidal transformation}

\item{scaled}{are data scaled ?}
}
\value{
Normalized data
}
\description{
Flatten Z-score by sigmoidal transformation
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
