% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorMatrix.R
\name{cor.matrix}
\alias{cor.matrix}
\title{Correlation matrix}
\usage{
cor.matrix(R, histogram = TRUE, method = c("pearson", "kendall",
  "spearman"), ...)
}
\arguments{
\item{R}{data set}

\item{histogram}{shall histogram be depicted on diagonal?}

\item{method}{method argument from cor function}
}
\value{
a correlation matrix plot
}
\description{
Slightly modified version of Chart.Correlation from PerformanceAnalytics package
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
