% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReduc.R
\name{data.reduc}
\alias{data.reduc}
\title{Basic statistics}
\usage{
data.reduc(data.table, NameIndex, VarIndexes, Ctrl, it = 0, spt = "_",
  sem = F, trunc.set = F, d.norm = F, alt = "two.sided")
}
\arguments{
\item{data.table}{A data.frame on which stats are applied}

\item{NameIndex}{The column name/index from data.table where condition name is found}

\item{VarIndexes}{The column names/indexes whit variables of interest}

\item{Ctrl}{A character string that contains the name of the control against which t.test will be performed}

\item{it}{Shall t.test/normalization be applied recursively ?}

\item{spt}{The linker used in condition names (see example)}

\item{sem}{Shall sem be calculated instead of sd ?}

\item{trunc.set}{Shall data outlyers be removed automatically?}

\item{d.norm}{Shall normalization to control be applied ?}

\item{alt}{See alt parameter from t.test function.}
}
\value{
A data.frame containing mean,sd/sem,p-value from t-test, asterisks of signativity for each variable
}
\description{
Calculate basic statistics from well-by-well data to condition data
}
\examples{
#Not running
Results = data.reduc(data.table = Summary,NameIndex=1,VarIndexes=10:13,
Ctrl='Ctr_8H',sem=T, trunc.set=T, d.norm=T, alt='greater')

}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
