% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindFocus.R
\name{FindFocus}
\alias{FindFocus}
\title{Out of focus detection}
\usage{
FindFocus(SERVER, PlateID, nCores = detectCores())
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database}

\item{PlateID}{the Plate Identifier}

\item{nCores}{the number of cores for parallel processing}
}
\value{
A list containing possible out ouf focus sites for each well
}
\description{
Return potential out of focus sites from the plate
}
\examples{
#Not running
OF = FindFocus(SERVER = 'MDCStore',ID = 'sa', PWD = '***') 

}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
