% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCELL.R
\name{GetCELL}
\alias{GetCELL}
\title{Cell Image Query}
\usage{
GetCELL(SERVER = "MDCStore", ID = "moldev", PWD = "moldev", PlateID,
  WellID, SiteID = c(1, 1), Range = lapply(1:3, function(x) c(0, 0.1)),
  use.autoRange = T, col.order = c("blue", "green", "red"),
  channel.choice = 1:3, disp = "raster", TimeCourse = F,
  TimePoint = 1, returnMat = F, CellStep = 50, xcoord, ycoord,
  fullIm = T, ...)
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database. Ignored if DB given}

\item{ID}{the identifiant for connexion. Ignored if DB given}

\item{PWD}{the password for connexion. Ignored if DB given}

\item{PlateID}{the PlateID of the image}

\item{WellID}{the WellID of the image}

\item{SiteID}{the SiteID of the image}

\item{Range}{a list containing range values dor each channel image normalization (see normalize from EBImage)}

\item{col.order}{the order of colours to be applied when combining channels}

\item{channel.choice}{channel selection (3 maximum)}

\item{disp}{'raster' or 'browser'. Where should be the image plotted. See display from EBImage package}

\item{TimeCourse}{does the image come from timecourse experiment ?}

\item{TimePoint}{the timepoint of the image. Ignored if TimeCourse=FALSE}

\item{returnMat}{shall image matrix be returned instead of dispayed ?}

\item{CellStep}{radius of the circle around from cell center coordinates.}

\item{xcoord}{x coordinate of cell pixel centroid}

\item{ycoord}{y coordinate of cell pixel centroid}

\item{fullIm}{shall full image be returned or cell zoom only?}

\item{resize}{shall the returned image be resized ?}

\item{dim.resize}{the size of resized image. Ignored if resize=FALSE}
}
\value{
a raster image or Image class element
}
\description{
Return the combined image of given plate location and mark cell coordinates with a white circle
}
\examples{
#Not running
#Bi-parametric plot of cell properties
smoothScatter(log10(GlobalData$Hoechst.Norm),log10(GlobalData$Nuc.Corr),colramp=myRamp,useRaster=T,
nrpoints=0,bandwidth=0.001,nbin=500,ylab='Nuclear Area',xlab='Corrected Hoechst Intensity',main='Debris Segmentation')

#Get coordinates of given point
test = identify(log10(GlobalData$Hoechst.Norm),log10(GlobalData$Nuc.Corr),n=1,plot=F)

#Draw image where cell is found
GetCELL(PlateID = GlobalData$Plate.ID[test],SiteID = GlobalData$Site.ID[test],WellID = GlobalData$Well.ID[test],xcoord = GlobalData$'Pixel Centroid X'[test],ycoord = GlobalData$'Pixel Centroid Y'[test],
           SERVER =SERVER,Range=list(c(0,0.4),c(0,0.1),c(0,0.05)),disp='browser',CellStep = 50, fullIm = T) #Return cell located on plot

}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
