% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIM.R
\name{GetIM}
\alias{GetIM}
\title{Whole Image Query}
\usage{
GetIM(SERVER = "MDCStore", ID = "moldev", PWD = "moldev", PlateID,
  WellID, SiteID = c(1, 1), Z = 1, Range = lapply(1:3, function(x)
  c(0, 0.1)), use.autoRange = T, col.order = c("blue", "green", "red"),
  channel.choice = 1:3, disp = "raster", TimeCourse = F,
  TimePoint = 1, resize = F, dim.resize = 600, returnMat = F, ...)
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database. Ignored if DB given}

\item{ID}{the identifiant for connexion. Ignored if DB given}

\item{PWD}{the password for connexion. Ignored if DB given}

\item{PlateID}{the PlateID of the image}

\item{WellID}{the WellID of the image}

\item{SiteID}{the SiteID of the image}

\item{Z}{The Z position of the image. Default is 1, value if no stacks is recorded. 0 for max projection, if recorded}

\item{Range}{a list containing range values dor each channel image normalization (see normalize from EBImage)}

\item{col.order}{the order of colours to be applied when combining channels}

\item{channel.choice}{channel selection (3 maximum)}

\item{disp}{'raster' or 'browser'. Where should be the image plotted. See display from EBImage package}

\item{TimeCourse}{does the image come from timecourse experiment ?}

\item{TimePoint}{the timepoint of the image. Ignored if TimeCourse=FALSE}

\item{resize}{shall the returned image be resized ?}

\item{dim.resize}{the size of resized image. Ignored if resize=FALSE}

\item{returnMat}{shall image matrix be returned instead of dispayed ?}
}
\value{
a raster image or EBImage::Image class element
}
\description{
Return the combined image of given plate location
}
\examples{
#Not running
GetIM(SERVER='YODA-SERVER', PlateID=1,WellID='E05',SiteID=c(2,2),disp='raster')

}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
