% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIMPath.R
\name{GetIMPath}
\alias{GetIMPath}
\title{Quick image name finding}
\usage{
GetIMPath(
  SERVER = "MDCStore",
  ID = "moldev",
  PWD = "moldev",
  PlateID,
  WellID,
  SiteID,
  TimePoint = 1,
  Z = 1,
  rdif = c("//", "//")
)
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database}

\item{ID}{the identifiant for connexion}

\item{PWD}{the password for connexion}

\item{PlateID}{the Plate Identifier}

\item{WellID}{the Well Identifier}

\item{SiteID}{a vector of 2 integers showing x and y site positions. If only one site present, x and y equal 0}

\item{TimePoint}{From what timepoint the image name should be returned. Default is 1.}

\item{Z}{The Z position of the image. Default is 1, value if no stacks is recorded. 0 for max projection, if recorded}

\item{rdif}{the string replacement for the mounting point of UNC image location}
}
\value{
A string array containing full image names
}
\description{
Get MX image name(s) given its coordinates
}
\examples{
#Not running
ips = GetIMPath(SERVER = 'MDCStore',ID = 'sa', PWD = '***',PlateID=10, WellID = 'A02', SiteID=c(1,1), TimePoint=1, Z=1) 
imgs = readImage(ips)

}
\seealso{
\code{\link{GetMDCInfo}}
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
