% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMDCData.R
\name{GetMDCData}
\alias{GetMDCData}
\title{Cell-by-Cell data Query}
\usage{
GetMDCData(SERVER = "MDCStore", ID = "moldev", PWD = "moldev",
  MeasurementID, getsite = T, TimeCourse = F)
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database}

\item{ID}{the identifiant for connexion}

\item{PWD}{the password for connexion}

\item{MeasurementID}{the uniqueID created by MetaXpress/PowerCore segmentation}

\item{getsite}{Shall siteID be added to data set ?}

\item{TimeCourse}{Is it a timecourse experiment?}
}
\value{
A data.frame with cell-by-cell data stemming from MetaXpress/PowerCore segmentation
}
\description{
Get Cell-by-Cell segmentation data stored in MDCStore database given uniqueID
}
\examples{
#Not running
DB = GetMDCInfo(SERVER = 'MDCStore',ID = 'sa', PWD = '***') 
UniqueIDs=DB$MeasurementID[grep(paste0(1500:1515,collapse='|'),DB$PlateID)]) #Get uniquesIDs from plate 1500 to 1515
MyData = lapply(UniqueIDs,function(x)GetMDCData(SERVER='YODA-SERVER',MeasurementID=x)) #Create a list that contains cellular data from each plate

}
\seealso{
\code{\link{GetMDCInfo}}
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
