% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPINFO.R
\name{GetPINFO}
\alias{GetPINFO}
\title{Retrieve plate information}
\usage{
GetPINFO(SERVER = "MDCStore", ID = "moldev", PWD = "moldev", PlateID)
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database}

\item{ID}{the identifiant for connexion}

\item{PWD}{the password for connexion}

\item{PlateID}{the Plate Identifier}
}
\value{
Plate information as a list
}
\description{
Get MX plate layout and series information
}
\examples{
#Not running
DB = GetPINFO(SERVER = 'MDCStore',ID = 'sa', PWD = '***',PLateID=1) 

}
\seealso{
\code{\link{GetMDCInfo}}
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
