% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHTD.R
\name{HTD.create}
\alias{HTD.create}
\title{HTD file creation}
\usage{
HTD.create(Pty = 1, WellList = c("A01", "A02"), TimePt = 1,
  PlateName = "Template", export, Sites = c(3, 3),
  Channels = c("DAPI", "GFP"))
}
\arguments{
\item{Pty}{1 for 384-well plate, 2 for 96-well plate}

\item{WellList}{List of wells "acquired" in plate}

\item{TimePt}{Timepoint number}

\item{PlateName}{Plate name}

\item{export}{Where to write the HTD file}

\item{Sites}{Number of sites along x and y dimensions}

\item{Channels}{MX channels}
}
\value{
An HTD file for plate import
}
\description{
Function that create the HTD file necessary to import images into MX database
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
