% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MX2FCS.R
\name{MX2FCS}
\alias{MX2FCS}
\title{Export cell data as Flow Cytometry Standard}
\usage{
MX2FCS(dat, coi, export, wlab = "Well.ID", proj = "")
}
\arguments{
\item{dat}{data frame coming from \code{\link{GetMDCData}}}

\item{coi}{column Of Interest to be exported into FCS}

\item{export}{folder location where data will be exported}

\item{wlab}{column name that contain well identifier. Default ix 'Well.ID' which is return by \code{\link{GetMDCData}}}

\item{proj}{character string with project name}
}
\description{
Retrieve cell data from database and export as Flow Cytometry Standard
}
\examples{
#Not running
mydat=GetMDCData(SERVER='MDCStore',MeasurementID=100)
MX2FCS(mydat,coi=c('Hoechst_Intensity','Nuclear_area'),export='MYFCS')

}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
