% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{Normalize}
\alias{Normalize}
\title{Image normalization}
\usage{
Normalize(ImObj, inputRange, autoRange = F, step = 0.005,
  verbose = T)
}
\arguments{
\item{ImObj}{An array, a matrix or an image}

\item{inputRange}{The targeted min and max pixel intensities}

\item{autoRange}{A boolean indicating if inputRange should be calculated automatically}

\item{step}{A numeric value, indicating probs from quantile option for inputRange automated calculation}

\item{verbose}{logical. Shall normalization range be displayed?}
}
\value{
An array, a matrix or an image with values ranged from 0 to 1
}
\description{
Normalize image pixel intensities between 0 and 1
}
\examples{
Mat = replicate(100, rnorm(100))
Mat.norm = Normalize(Mat);Mat.autoNorm = Normalize(Mat,autoRange=T)
plot(density(Mat));
plot(density(Mat.norm));points(density(Mat.autoNorm),type='l',col='blue')


}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
