% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakCalc.R
\name{peak.calc}
\alias{peak.calc}
\title{Better than mean for non-Gaussian distribution scaling}
\usage{
peak.calc(x, use.log = F)
}
\arguments{
\item{x}{A numerical vector}

\item{use.log}{Shall a log10 transformation be applied ?}
}
\value{
A double which corresponds to the abciss of maximum density
}
\description{
Get peak value of data distribution
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
