% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCM.R
\name{SaveCM}
\alias{SaveCM}
\title{Retrieve plate information}
\usage{
SaveCM(
  SERVER = "MDCStore",
  ID = "moldev",
  PWD = "moldev",
  IR = 1,
  exp_dir = getwd()
)
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database}

\item{ID}{the identifiant for connexion}

\item{PWD}{the password for connexion}

\item{IR}{the number of rows at a time to fetch}

\item{exp_dir}{where to export xml files}
}
\value{
Plate information as a list
}
\description{
Export all Custom modules from database as xml files
}
\examples{
#Not running
SaveCM(SERVER = 'MDCStore',ID = 'sa', PWD = '***') 

}
\seealso{
\code{\link{GetMDCInfo}}
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
