% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatMode.R
\name{stat.mode}
\alias{stat.mode}
\title{Better than mean for non-Gaussian distribution scaling}
\usage{
stat.mode(x, use.log = F)
}
\arguments{
\item{x}{A numerical vector}

\item{use.log}{Shall a log10 transformation be applied ?}
}
\value{
A double which corresponds to the max value of the probability mass function
}
\description{
Get statistical mode of data distribution
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
