% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIMPath.R
\name{GetIMPath}
\alias{GetIMPath}
\title{Quick image name finding}
\usage{
GetIMPath(SERVER = "MDCStore", ID = "moldev", PWD = "moldev",
  PlateID, WellID, SiteID, TimePoint = 1, Z = 1, Unix.diff = c("//",
  "/media/"))
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database}

\item{ID}{the identifiant for connexion}

\item{PWD}{the password for connexion}

\item{PlateID}{the Plate Identifier}

\item{WellID}{the Well Identifier}

\item{SiteID}{a vector of 2 integers showing x and y site positions. If only one site present, x and y equal 0}

\item{TimePoint}{From what timepoint the image name should be returned. Default is 1.}

\item{Z}{The Z position of the image. Default is 1, value if no stacks is recorded. 0 for max projection, if recorded}

\item{Unix.diff}{the string replacement for the mounting point of UNC image location, when accessed from linux machine}
}
\value{
A string array containing full image names
}
\description{
Get MX image name(s) given its coordinates
}
\examples{
#Not running
DB = GetMDCInfo(SERVER = 'MDCStore',ID = 'sa', PWD = '***') 
UniqueIDs=DB$MeasurementID[grep(paste0(1500:1515,collapse='|'),DB$PlateID)]) #Get uniquesIDs from plate 1500 to 1515
MyData = lapply(UniqueIDs,function(x)GetMDCData(SERVER='YODA-SERVER',MeasurementID=x)) #Create a list that contains cellular data from each plate

}
\seealso{
\code{\link{GetMDCInfo}}
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
