% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMDCInfo.R
\name{GetMDCInfo}
\alias{GetMDCInfo}
\title{MDCStore information Query}
\usage{
GetMDCInfo(SERVER = "MDCStore", ID = "moldev", PWD = "moldev",
  Unix.diff = c("//", "/media/"))
}
\arguments{
\item{SERVER}{the odbc connexion to MDCStore database}

\item{ID}{the identifiant for connexion}

\item{PWD}{the password for connexion}

\item{Unix.diff}{the string replacement for the mounting point of UNC image location, when accessed from linux machine}
}
\value{
a data.frame with set of information about acquired plates
}
\description{
Retrieve plate informations stored in MDCStore database
}
\examples{
#Not running
DB = GetMDCInfo(SERVER = 'MDCStore',ID = 'sa', PWD = '***') 
View(DB[DB$PlateID==1,]) #Get information about plate 1

}
\seealso{
\code{\link{GetMDCData}}
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
\keyword{utilities}
